Move test added in b6c89b9a32a to ..._cipherstash.rs to avoid losing it in the upgrade.
