+++ packages/sqltk-parser/tests/sqlparser_postgres_cipherstash.rs
@@ -23,6 +23,7 @@
 mod test_utils;
 use test_utils::*;
 
+use sqltk_parser::ast::*;
 use sqltk_parser::dialect::PostgreSqlDialect;
 
 // Helpers
@@ -66,4 +67,34 @@ fn parse_lock_table() {
     pg().verified_stmt("LOCK TABLE ONLY customers, orders IN EXCLUSIVE MODE");
     pg().verified_stmt("LOCK TABLE ONLY customers, orders IN ACCESS EXCLUSIVE MODE");
     pg().verified_stmt("LOCK TABLE ONLY customers, orders IN ACCESS SHARE MODE NOWAIT");
+}
+
+#[test]
+fn parse_arbitrary_operator_with_any_and_all() {
+    // The following statement is semantically non-sensical on vanilla postgres but it should *parse* in order to
+    // support situations where PG operators have been overloaded.
+    let select = pg().verified_only_select("SELECT 123 % ANY(ARRAY[1, 2, 3])");
+    assert!(
+        matches!(
+            select.projection[0],
+            SelectItem::UnnamedExpr(Expr::AnyOp {
+                left: _,
+                compare_op: BinaryOperator::Modulo,
+                right: _,
+                is_some: false
+            })
+        )
+    );
+
+    let select = pg().verified_only_select("SELECT 123 % ALL(ARRAY[1, 2, 3])");
+    assert!(
+        matches!(
+            select.projection[0],
+            SelectItem::UnnamedExpr(Expr::AllOp {
+                left: _,
+                compare_op: BinaryOperator::Modulo,
+                right: _,
+            })
+        )
+    );
 }
